fix: dev as default branch
