+++ targets/export-elasticsearch/package.json
@@ -33,6 +33,7 @@
     "zod": "^3.14.2"
   },
   "devDependencies": {
+    "@shared/types": "^1.14.2",
     "@swc/cli": "0.1.55",
     "@swc/core": "1.2.150",
     "@swc/jest": "0.2.20",
