+++ targets/export-elasticsearch/src/controllers/__test__/fake/export.ts
@@ -1,7 +1,7 @@
+import type { ExportEsStatus } from "@shared/types";
+import { Environment, Status } from "@shared/types";
 import { injectable } from "inversify";
 
-import type { ExportEsStatus } from "../../../types";
-import { Environment, Status } from "../../../types";
 import { wait } from "../../../utils";
 
 @injectable()
