--- targets/export-elasticsearch/src/types.ts
@@ -1,28 +1,0 @@
-export enum Status {
-  running = "running",
-  completed = "completed",
-  failed = "failed",
-  timeout = "timeout",
-}
-
-export enum Environment {
-  production = "production",
-  preproduction = "preproduction",
-}
-
-export interface User {
-  id: string;
-  name: string;
-  email: string;
-  created_at: Date;
-}
-
-export interface ExportEsStatus {
-  id: string;
-  environment: Environment;
-  status: Status;
-  user_id: string;
-  created_at: Date;
-  updated_at: Date;
-  user?: User;
-}
