+++ targets/frontend/package.json
@@ -76,8 +76,7 @@
     "unist-util-select": "^4.0.1",
     "urql": "^2.0.5",
     "uuid": "^8.3.2",
-    "wonka": "^4.0.15",
-    "zustand": "^3.7.1"
+    "wonka": "^4.0.15"
   },
   "devDependencies": {
     "@shared/types": "^1.14.2",
