+++ targets/frontend/src/components/export-es/Status.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource theme-ui */
 
+import { Status as StatusType } from "@shared/types";
 import { MdTimelapse } from "react-icons/md";
-import { Status as StatusType } from "src/store/export-es";
 import { Box, Flex } from "theme-ui";
 
 type StatusProps = {
